Upgrading/adapting to osmdroid 4.2

+++ osmdroid-android/src/org/osmdroid/bonuspack/overlays/MapEventsOverlay.java
@@ -2,7 +2,7 @@ package org.osmdroid.bonuspack.overlays;
 
 import org.osmdroid.util.GeoPoint;
 import org.osmdroid.views.MapView;
-import org.osmdroid.views.MapView.Projection;
+import org.osmdroid.views.Projection;
 import org.osmdroid.views.overlay.Overlay;
 import android.content.Context;
 import android.graphics.Canvas;
@@ -34,13 +34,13 @@ public class MapEventsOverlay extends Overlay {
 	
 	@Override public boolean onSingleTapConfirmed(MotionEvent e, MapView mapView){
 		Projection proj = mapView.getProjection();
-		GeoPoint p = (GeoPoint)proj.fromPixels(e.getX(), e.getY());
+		GeoPoint p = (GeoPoint)proj.fromPixels((int)e.getX(), (int)e.getY());
 		return mReceiver.singleTapConfirmedHelper(p);
 	}
 	
 	@Override public boolean onLongPress(MotionEvent e, MapView mapView) {
 		Projection proj = mapView.getProjection();
-		GeoPoint p = (GeoPoint)proj.fromPixels(e.getX(), e.getY());
+		GeoPoint p = (GeoPoint)proj.fromPixels((int)e.getX(), (int)e.getY());
 		//throw event to the receiver:
 		return mReceiver.longPressHelper(p);
     }
